chore(vue-flow,store): move store utils to separate file
