+++ packages/vue-flow/src/utils/index.ts
@@ -2,3 +2,4 @@ export * from './edge'
 export * from './graph'
 export * from './node'
 export * from './changes'
+export * from './store'
